tools/gemini_service/requirements.txt
flask
google-generativeai
requests
